Added number type parameter loading + bugfixes
- Added loading of number type parameters to the edit window
- Fixed a bug where the add image button would not show up in the edit window
- Fixed a bug where you could not click on the arrows in photo view if the image was too big
- Fixed the detail window not showing any images
- Fixed crashes of the app when clicking on the description buttons while it is empty
- Fixed the edit window duplicating component parameters on saving
- Fixed some more edge-case scenarios crashes with protections and checks

+++ Pc parts lister/Edit_window.xaml.cs
@@ -43,7 +43,7 @@ namespace Pc_parts_lister
 
             MainPicButton.Click += OpenImage_Click;
 
-            if (komponenta.FullImagePath != null || komponenta.FullImagePath != "")
+            if (komponenta.FullImagePath != null && komponenta.FullImagePath != "")
             {
                 SelectMainPic_Button.Visibility = Visibility.Collapsed;
                 MainPicButton.Height = 250;
@@ -189,6 +189,16 @@ namespace Pc_parts_lister
                 box.Tag = parameter;
                 box.SelectionChanged += Check_Int_Validation;
 
+                bool found = false;
+                foreach (Parameter kparameter in Komponenta.parameters)
+                {
+                    if (kparameter.ID == parameter.ID)
+                    {
+                        found = true;
+                        box.Text = kparameter.Value;
+                    }
+                }
+
                 panel.Children.Add(box);
 
                 return panel;
@@ -244,8 +254,10 @@ namespace Pc_parts_lister
         {
             Close();
         }
+
         private void Save_Click(object sender, RoutedEventArgs e)
         {
+            Komponenta.parameters.Clear();
             foreach (StackPanel panel in Parameters_Panel.Children)
             {
                 PossibleParameter parameter = (PossibleParameter)panel.Tag;
@@ -443,8 +455,11 @@ namespace Pc_parts_lister
             OpenFileDialog dialog = new OpenFileDialog();
             dialog.Filter = "Obrázky (*.png;*.jpg;*.jpeg)|*.png;*.jpg;*.jpeg";
 
-            File.Delete(Komponenta.FullImagePath);
-
+            if (Komponenta.FullImagePath != null)
+            {
+                File.Delete(Komponenta.FullImagePath);
+            }
+            
             if (dialog.ShowDialog() == true)
             {
                 string imagesFolder = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "Images");
